README, funkcny AWGN kanal a divny BER

+++ Projekt/gr-ErTools/lib/BER_impl.h
@@ -17,9 +17,9 @@ class BER_impl : public BER
 {
 private:
   int _N;
-  std::vector<unsigned long long> _count;
+  std::vector<int> _count;
   std::vector<double> _pamat_SER;
-  std::vector<unsigned long long> _pocet_chyb;
+  std::vector<int> _pocet_chyb;
   /*int _count[128], _pocet_chyb[128]; 
   float _pamat_SER[128];
 */
